fix(portal): production error from sidebar rendering server side
