+++ libs/portal/pages/index/index.tsx
@@ -5,6 +5,7 @@ import { Stat } from '@cuhacking/portal/ui/dashboard/stat'
 import { UserStatus } from '@cuhacking/portal/ui/dashboard/user-status'
 import { Sidebar } from '@cuhacking/portal/ui/sidebar'
 import { Button } from '@cuhacking/shared/ui/button'
+import { ClientOnly } from 'remix-utils/client-only'
 
 const constants = {
   user: {
@@ -23,7 +24,10 @@ const constants = {
 export function Home() {
   return (
     <div className="flex h-screen">
-      <Sidebar />
+      <ClientOnly>
+        {() =>
+          (<Sidebar />)}
+      </ClientOnly>
       <div className="flex flex-1 flex-col w-full p-4 gap-5">
         <Banner name={constants.user.name} />
         <div className="flex justify-center w-full">
